requirements.txt
PyYAML
gitpython
pytest-mock
pytest
Jinja2
PyInstaller
requests